mc_microservices/app/requirements.txt
fastapi
uvicorn
pydantic
httpx
python-multipart
python-dotenv
apscheduler
hvac